@@ -12,17 +12,17 @@
   <packaging>jar</packaging>
 
   <properties>
-    <beam.version>2.47.0</beam.version>
+    <beam.version>2.49.0</beam.version>
 
     <freemarker.version>2.3.32</freemarker.version>
     <hadoop.version>3.3.5</hadoop.version>
     <hbase.version>2.5.4</hbase.version>
     <onnxruntime.version>1.15.1</onnxruntime.version>
     <onnx.version>1.14.0-1.5.9</onnx.version>
-    <solr.version>9.2.1</solr.version>
-    <lucene.version>9.4.2</lucene.version>
-    <neo4j.version>4.4.21</neo4j.version>
-    <pdfbox.version>2.0.28</pdfbox.version>
+    <solr.version>9.3.0</solr.version>
+    <lucene.version>9.7.0</lucene.version>
+    <neo4j.version>5.11.0</neo4j.version>
+    <pdfbox.version>2.0.29</pdfbox.version>
     <jts.version>1.19.0</jts.version>
     <mysql-connector.version>8.0.33</mysql-connector.version>
     <postgresql-connector.version>42.6.0</postgresql-connector.version>
@@ -64,8 +64,8 @@
         <artifactId>maven-compiler-plugin</artifactId>
         <version>${maven-compiler-plugin.version}</version>
         <configuration>
-          <source>11</source>
-          <target>11</target>
+          <source>17</source>
+          <target>17</target>
         </configuration>
       </plugin>
 
@@ -143,7 +143,7 @@
         </executions>
         <configuration>
           <from>
-            <image>gcr.io/dataflow-templates-base/java11-template-launcher-base:latest</image>
+            <image>gcr.io/dataflow-templates-base/java17-template-launcher-base:latest</image>
           </from>
           <container>
             <appRoot>/template/FlexPipeline</appRoot>
@@ -247,10 +247,10 @@
       <version>${beam.version}</version>
     </dependency>
 
-    <!-- https://mvnrepository.com/artifact/org.apache.beam/beam-sdks-java-io-amazon-web-services -->
+    <!-- https://mvnrepository.com/artifact/org.apache.beam/beam-sdks-java-io-amazon-web-services2 -->
     <dependency>
       <groupId>org.apache.beam</groupId>
-      <artifactId>beam-sdks-java-io-amazon-web-services</artifactId>
+      <artifactId>beam-sdks-java-io-amazon-web-services2</artifactId>
       <version>${beam.version}</version>
     </dependency>
 
